requirements.txt
Flask
flask-sqlalchemy
flask-migrate
flask-jwt-extended
psycopg2-binary
python-dotenv
flask-cors
werkzeug
email-validator